Ads fixed.
Used some of Newb420's fix. I'm Planning a large update but it's delayed after losing my laptops data.
New GUI is planned, I want to make it easier to change units on the fly and remove the configs. Right now green maps are used automatically but I will make it optional along with the use of gold maps.
Any changes that you would like to see please add them to issues and I'll see if they are viable.